backend/requirements.txt
# Contenuto per requirements.txt

# Core del Web Framework
Flask~=3.1.1
Flask-SQLAlchemy~=3.1.1
Flask-Migrate~=4.1.0
Flask-Login~=0.6.3
Flask-Cors~=6.0.0
Flask-Limiter~=3.5.0  # Rate limiting

# Gestione Database
SQLAlchemy~=2.0.41
alembic~=1.16.5

# Cache & Performance
redis~=5.0.1  # Cache e rate limiting storage

# Task Queue
celery~=5.3.4  # Background tasks
celery[redis]~=5.3.4  # Redis broker per Celery

# Logica di Business
stripe~=13.0.1
Authlib~=1.6.4
Pillow~=11.1.0  # Per la gestione delle immagini

# Utilità
python-dotenv  # Per caricare il file .env

# Production WSGI server
gunicorn~=23.0.0
